Deletes .metadata/ and updates .gitignore
To ensure .metadata/ is not retained in future iterations, its file was
completely deleted to reset the tracking of .gitignore files. .gitignore
has now been verified to filter all files not directly related to the
project for cleaner updates, and as such should be the new standard in
all iterations.