Partially Implements Game Scene View w/ Controller
Game Scene within GameScene.fxml is mostly constructed, with proper
grouping in place and ToggleGroup (the equivalent of a RadioGroup in
Swing) being implemented for each of the Radio Buttons appropriately.
Everything has been resized appropriately to appear reasonably centered,
and is slowly inching itself to a decent game state position (suitable
for Sprint 2 at minimum).